fixed word long long

+++ gameImplementation.h
@@ -21,12 +21,12 @@ enum RETURNVALUESGAMERUNTIME{
 int gameRuntime(char *activeWord);
 
 
-void printVariablyCoveredWord(unsigned long long wordSize,const short int *uncoveredArray,
+void printVariablyCoveredWord(unsigned short wordSize,const short int *uncoveredArray,
                               char *activeWord, char *statusWordUncovered);
 
 short int coveredWordManagement(char inputChar, char *convertedWord, short int *uncoveredArray,
                                 char *misses, char *StatusWordForDataLogging);
 
 
-short int sequenceLetUserGuessWord(char * activeWordConvert, unsigned long long activeWordLength);
+short int sequenceLetUserGuessWord(char * activeWordConvert, unsigned short activeWordLength);
